+++ changelog.md
@@ -8,6 +8,7 @@
 - Add missing stamp recipes to Carpenter (#228)
 - Fix Vanilla fireproof log tags being empty (#237)
 - Fix worldgen crash caused by Undermod biomes (#229)
+- Add config options to forestry-server.toml to disable Apatite and Tin generation (#108) 
 
 ## Forestry 2.6.0
 - Fixed planks recipes for all wood types, now able to craft planks from stripped logs (#227)